requirements.txt
pdfminer.six
pdfplumber